requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
gtts==2.5.1
python-dotenv==1.0.0
requests==2.31.0